Serial C radial classifier

+++ src/cudaDecompose.h
@@ -1,6 +1,6 @@
 #ifndef CUDA_DECOMPOSE
 #define CUDA_DECOMPOSE
 
-void RadialMean(int steps, int h, int w,  float *image, long *classified, float *means);
+void RadialMean(int steps, int h, int w,  float *image, float *classified, float *means);
 
 #endif
